# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lingmo",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/novels/{novel_id}/agent-report": {
      "get": {
        "tags": [
          "agent-pipeline"
        ],
        "summary": "Agent Report",
        "description": "Return all agent results from the last pipeline run for this novel.",
        "operationId": "agent_report_api_novels__novel_id__agent_report_get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Agent Report Api Novels  Novel Id  Agent Report Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{novel_id}/agent-pipeline": {
      "post": {
        "tags": [
          "agent-pipeline"
        ],
        "summary": "Run Agent Pipeline",
        "description": "Run the complete Agent pipeline for next chapter generation.",
        "operationId": "run_agent_pipeline_api_novels__novel_id__agent_pipeline_post",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Novel Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "default": {},
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Run Agent Pipeline Api Novels  Novel Id  Agent Pipeline Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/novels/{novel_id}/preview-constraints": {
      "get": {
        "tags": [
          "constraints"
        ],
        "summary": "Preview Constraints",
        "description": "Preview constraints that will be injected into next chapter generation.",
        "operationId": "preview_constraints_api_novels__novel_id__preview_constraints_get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Novel Id"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "L1",
              "title": "Level"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Preview Constraints Api Novels  Novel Id  Preview Constraints Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{novel_id}/test-constraints": {
      "get": {
        "tags": [
          "constraints"
        ],
        "summary": "Test Constraint Compression",
        "description": "A/B test: compare all 4 constraint compression levels.",
        "operationId": "test_constraint_compression_api_novels__novel_id__test_constraints_get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Test Constraint Compression Api Novels  Novel Id  Test Constraints Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/novels/{novel_id}/constraint-collapse": {
      "post": {
        "tags": [
          "constraints"
        ],
        "summary": "Constraint Collapse",
        "description": "Narrow scene choices through hard, character, structure, and theme constraints.",
        "operationId": "constraint_collapse_api_novels__novel_id__constraint_collapse_post",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Novel Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Constraint Collapse Api Novels  Novel Id  Constraint Collapse Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/novels/{novel_id}/draft": {
      "post": {
        "tags": [
          "drafts"
        ],
        "summary": "Draft Directions",
        "operationId": "draft_directions_api_novels__novel_id__draft_post",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Novel Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Draft Directions Api Novels  Novel Id  Draft Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{novel_id}/expand": {
      "post": {
        "tags": [
          "drafts"
        ],
        "summary": "Expand Chapter",
        "operationId": "expand_chapter_api_novels__novel_id__expand_post",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Novel Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Expand Chapter Api Novels  Novel Id  Expand Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{novel_id}/preview": {
      "get": {
        "tags": [
          "drafts"
        ],
        "summary": "Preview Chapter",
        "description": "Generate a short style preview without creating a full chapter.",
        "operationId": "preview_chapter_api_novels__novel_id__preview_get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Preview Chapter Api Novels  Novel Id  Preview Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}